refactor layers and clean up; hover DRY
